fixes and renaming
