+++ anvil/src/winit.rs
@@ -29,7 +29,7 @@ use smithay::{
     },
     utils::IsAlive,
     wayland::{
-        input_method_manager::InputMethodSeatTrait,
+        input_method::InputMethodSeat,
         output::{Mode, Output, PhysicalProperties},
         seat::CursorImageStatus,
     },
@@ -223,9 +223,7 @@ pub fn run_winit(log: Logger) {
             let input_method = state.seat.input_method();
             let (x, y, _, height) = input_method.coordinates();
             input_method.with_surface(|surface| {
-                if surface.alive() {
-                    elements.push(draw_input_popup_surface(surface.clone(), (x, (y + height))).into());
-                }
+                elements.push(draw_input_popup_surface(surface.clone(), (x, (y + height))).into());
             });
 
             // draw the cursor as relevant
